refactor(settings-form): extract form from add-data-source
