+++ settings-form/src/datasource/add-data-source/index.tsx
@@ -1,0 +1,98 @@
+import { Box, Button, Modal, SegmentedControl } from '@mantine/core';
+import { showNotification, updateNotification } from '@mantine/notifications';
+import React, { useState } from 'react';
+import { PlaylistAdd } from 'tabler-icons-react';
+import { APICaller } from '../../api-caller';
+import { DataSourceType } from '../../api-caller/datasource.typed';
+import { defaultStyles, IStyles } from '../styles';
+import { AddDataSourceForm_DB } from './forms/database';
+import { IFormValues } from './types';
+
+interface IAddDataSourceForm {
+  postSubmit: () => void;
+  styles?: IStyles;
+}
+
+function AddDataSourceForm({ postSubmit, styles = defaultStyles }: IAddDataSourceForm) {
+  const [type, setType] = useState<DataSourceType>('postgresql');
+  const addDataSource = async ({ type, key, config }: IFormValues) => {
+    showNotification({
+      id: 'for-creating',
+      title: 'Pending',
+      message: 'Adding data source...',
+      loading: true,
+    });
+    try {
+      console.log({ type, key, config });
+      // await APICaller.datasource.create(type, key, config);
+      updateNotification({
+        id: 'for-creating',
+        title: 'Successful',
+        message: 'Data source is added',
+        color: 'green',
+      });
+      postSubmit();
+    } catch (error: $TSFixMe) {
+      updateNotification({
+        id: 'for-creating',
+        title: 'Failed',
+        message: error.message,
+        color: 'red',
+      });
+    }
+  };
+
+  const isDBType = type === 'postgresql' || type === 'mysql';
+  return (
+    <Box mx="auto">
+      <SegmentedControl
+        fullWidth
+        mb={styles.spacing}
+        size={styles.size}
+        data={[
+          { label: 'PostgreSQL', value: 'postgresql' },
+          { label: 'MySQL', value: 'mysql' },
+          { label: 'HTTP', value: 'http' },
+        ]}
+        value={type}
+        onChange={(v: DataSourceType) => setType(v)}
+      />
+      {isDBType && <AddDataSourceForm_DB submit={addDataSource} styles={styles} type={type} />}
+    </Box>
+  );
+}
+
+interface IAddDataSource {
+  styles?: IStyles;
+  onSuccess: () => void;
+}
+
+export function AddDataSource({ onSuccess, styles = defaultStyles }: IAddDataSource) {
+  const [opened, setOpened] = React.useState(false);
+  const open = () => setOpened(true);
+  const close = () => setOpened(false);
+  const postSubmit = () => {
+    onSuccess();
+    close();
+  };
+
+  return (
+    <>
+      <Modal
+        overflow="inside"
+        opened={opened}
+        onClose={() => setOpened(false)}
+        title="Add a data source"
+        trapFocus
+        onDragStart={(e) => {
+          e.stopPropagation();
+        }}
+      >
+        <AddDataSourceForm postSubmit={postSubmit} styles={styles} />
+      </Modal>
+      <Button size={styles.button.size} onClick={open} leftIcon={<PlaylistAdd size={20} />}>
+        Add a Data Source
+      </Button>
+    </>
+  );
+}
