+++ settings-form/src/datasource/add-data-source/types.ts
@@ -1,0 +1,7 @@
+import { DataSourceType, TDataSourceConfig } from '../../api-caller/datasource.typed';
+
+export interface IFormValues {
+  type: DataSourceType;
+  key: string;
+  config: TDataSourceConfig;
+}
